requirements.txt
pyrogram
tgcrypto
requests
google-api-python-client
google-auth
google-auth-httplib2